refactor(cli): Centralize agent sync lifecycle
