+++ packages/cli/src/live-scan-store.test.ts
@@ -277,6 +277,7 @@ vi.mock("node:worker_threads", () => ({
 }));
 
 import { LiveScanStore, resolveAgentWatchTargets, type SessionsUpdatedEvent } from "./live-scan.js";
+import { AgentSyncEngine } from "./agent-sync-engine.js";
 import { appLogger } from "./logging.js";
 import { ThreadWorkerRunner, type WorkerResult, type WorkerRunner } from "./worker-runner.js";
 
@@ -286,6 +287,29 @@ function makeWorkerRunner() {
     run: vi.fn<WorkerRunner["run"]>(),
     shutdown: vi.fn<WorkerRunner["shutdown"]>(async () => undefined),
   } satisfies WorkerRunner;
+}
+
+function syncEngineOf(store: LiveScanStore): AgentSyncEngine {
+  return (store as unknown as { syncEngine: AgentSyncEngine }).syncEngine;
+}
+
+function runBackfill(store: LiveScanStore, agentName: string) {
+  const engine = syncEngineOf(store) as unknown as {
+    serialize: (
+      agentName: string,
+      kind: "backfill",
+      operation: () => Promise<unknown>,
+    ) => Promise<unknown>;
+    performBackfill: (agentName: string) => Promise<unknown>;
+  };
+  return engine.serialize(agentName, "backfill", () => engine.performBackfill(agentName));
+}
+
+function setRefreshDuration(store: LiveScanStore, agentName: string, durationMs: number): void {
+  const engine = syncEngineOf(store) as unknown as {
+    state: (agentName: string) => { lastRefreshDurationMs: number };
+  };
+  engine.state(agentName).lastRefreshDurationMs = durationMs;
 }
 
 let restoreRuntime: (() => void) | null = null;
@@ -468,7 +492,10 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false, { agents: ["codex", "kimi"] });
+    const store = new LiveScanStore({
+      watchEnabled: false,
+      scanOptions: { agents: ["codex", "kimi"] },
+    });
     await store.initialize();
 
     const snapshot = store.getSnapshot();
@@ -517,7 +544,7 @@ describe("LiveScanStore", () => {
       cacheTimestamps: { codex: 500 },
     });
 
-    const store = new LiveScanStore(false, {}, {}, { deferInitialRefresh: true });
+    const store = new LiveScanStore({ watchEnabled: false, deferInitialRefresh: true });
     const events: unknown[] = [];
     const statusEvents: unknown[] = [];
     store.subscribe((event) => events.push(event));
@@ -623,7 +650,11 @@ describe("LiveScanStore", () => {
           ),
     );
 
-    const store = new LiveScanStore(false, {}, { from: 3000 }, { deferInitialRefresh: true });
+    const store = new LiveScanStore({
+      watchEnabled: false,
+      startupScanOptions: { from: 3000 },
+      deferInitialRefresh: true,
+    });
     await store.initialize();
 
     store.startBackgroundRefresh();
@@ -673,7 +704,11 @@ describe("LiveScanStore", () => {
           ),
     );
 
-    const store = new LiveScanStore(false, {}, { from: 3000 }, { deferInitialRefresh: true });
+    const store = new LiveScanStore({
+      watchEnabled: false,
+      startupScanOptions: { from: 3000 },
+      deferInitialRefresh: true,
+    });
     await store.initialize();
 
     store.startBackgroundRefresh();
@@ -738,14 +773,14 @@ describe("LiveScanStore", () => {
           }),
       )
       .mockResolvedValueOnce({ sessions: [fresh], meta: {}, changedIds: [fresh.id] });
-    const store = new LiveScanStore(false, {}, {}, { workerRunner });
+    const store = new LiveScanStore({ watchEnabled: false, workerRunner });
     await store.initialize();
     const logInfo = vi.spyOn(appLogger, "info").mockImplementation(() => undefined);
 
-    const backfill = (store as any).runBackfill("codex");
+    const backfill = runBackfill(store, "codex");
     await vi.waitFor(() => expect(workerRunner.run).toHaveBeenCalledTimes(1));
 
-    const refresh = (store as any).refreshAgent("codex");
+    const refresh = syncEngineOf(store).refresh("codex");
     await Promise.resolve();
     expect(workerRunner.run).toHaveBeenCalledTimes(1);
 
@@ -789,11 +824,11 @@ describe("LiveScanStore", () => {
     workerRunner.run
       .mockRejectedValueOnce(new Error("backfill failed"))
       .mockResolvedValueOnce({ sessions: [fresh], meta: {}, changedIds: [fresh.id] });
-    const store = new LiveScanStore(false, {}, {}, { workerRunner });
+    const store = new LiveScanStore({ watchEnabled: false, workerRunner });
     await store.initialize();
 
-    const backfill = (store as any).runBackfill("codex");
-    const refresh = (store as any).refreshAgent("codex");
+    const backfill = runBackfill(store, "codex");
+    const refresh = syncEngineOf(store).refresh("codex");
     await Promise.all([backfill, refresh]);
 
     expect(consoleError).toHaveBeenCalledWith("[codex] Backfill failed:", expect.any(Error));
@@ -837,13 +872,13 @@ describe("LiveScanStore", () => {
           releaseBackfill = resolve;
         }),
     );
-    const store = new LiveScanStore(false, {}, {}, { workerRunner });
+    const store = new LiveScanStore({ watchEnabled: false, workerRunner });
     await store.initialize();
 
-    const backfill = (store as any).runBackfill("codex");
+    const backfill = runBackfill(store, "codex");
     await vi.waitFor(() => expect(workerRunner.run).toHaveBeenCalledTimes(1));
 
-    await (store as any).refreshAgent("kimi");
+    await syncEngineOf(store).refresh("kimi");
     expect(store.getSnapshot().byAgent.kimi![0]?.title).toBe("kimi fresh");
 
     releaseBackfill!({ sessions: [codexInitial], meta: {}, changedIds: [] });
@@ -882,23 +917,21 @@ describe("LiveScanStore", () => {
       )
       .mockResolvedValueOnce({ sessions: [refreshed], meta: {}, changedIds: [refreshed.id] })
       .mockResolvedValueOnce({ sessions: [rerun], meta: {}, changedIds: [rerun.id] });
-    const store = new LiveScanStore(false, {}, {}, { workerRunner });
+    const store = new LiveScanStore({ watchEnabled: false, workerRunner });
     await store.initialize();
 
-    const backfill = (store as any).runBackfill("codex");
+    const backfill = runBackfill(store, "codex");
     await vi.waitFor(() => expect(workerRunner.run).toHaveBeenCalledTimes(1));
 
-    (store as any).refreshes.recordChangedPaths("codex");
-    (store as any).scheduleRefresh("codex", 0);
-    await vi.advanceTimersByTimeAsync(0);
-    (store as any).refreshes.recordChangedPaths("codex", 2);
-    (store as any).scheduleRefresh("codex", 0);
-    (store as any).scheduleRefresh("codex", 0);
-    await vi.advanceTimersByTimeAsync(0);
+    const refresh = syncEngineOf(store).refresh("codex");
+    await Promise.resolve();
+    await syncEngineOf(store).refresh("codex");
+    await syncEngineOf(store).refresh("codex");
 
     expect(workerRunner.run).toHaveBeenCalledTimes(1);
     releaseBackfill!({ sessions: [backfilled], meta: {}, changedIds: [backfilled.id] });
     await backfill;
+    await refresh;
     await vi.waitFor(() => expect(workerRunner.run).toHaveBeenCalledTimes(2));
 
     await vi.advanceTimersByTimeAsync(100);
@@ -990,7 +1023,11 @@ describe("LiveScanStore", () => {
           ),
     );
 
-    const store = new LiveScanStore(false, {}, { from: 3000 }, { deferInitialRefresh: true });
+    const store = new LiveScanStore({
+      watchEnabled: false,
+      startupScanOptions: { from: 3000 },
+      deferInitialRefresh: true,
+    });
     const events: unknown[] = [];
     store.subscribe((event) => events.push(event));
     await store.initialize();
@@ -1059,11 +1096,11 @@ describe("LiveScanStore", () => {
       timestamp: 1000,
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     const events: unknown[] = [];
     store.subscribe((event) => events.push(event));
     await store.initialize();
-    await (store as any).runRefresh("codex");
+    await syncEngineOf(store).refresh("codex");
     await vi.advanceTimersByTimeAsync(250);
 
     expect(codex.checkForChanges).not.toHaveBeenCalled();
@@ -1158,9 +1195,9 @@ describe("LiveScanStore", () => {
       timestamp: 1000,
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
-    await (store as any).runRefresh("codex");
+    await syncEngineOf(store).refresh("codex");
 
     expect(scanSessionSource).not.toHaveBeenCalled();
     expect(workerThreads.workers.at(-1)?.workerData.jobs).toEqual([
@@ -1202,11 +1239,11 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     const events: unknown[] = [];
     store.subscribe((event) => events.push(event));
     await store.initialize();
-    await (store as any).runRefresh("codex");
+    await syncEngineOf(store).refresh("codex");
     await vi.advanceTimersByTimeAsync(250);
 
     expect(codex.checkForChanges).toHaveBeenCalledWith(expect.any(Number), [previous]);
@@ -1268,10 +1305,10 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
-    (store as any).refreshes.recordChangedPaths("codex", 101);
-    await (store as any).runRefresh("codex");
+    syncEngineOf(store).handleAgentsChanged(Array.from({ length: 101 }, () => "codex"));
+    await syncEngineOf(store).refresh("codex");
 
     expect(workerThreads.workers.at(-1)?.workerData.jobs[0]).toEqual(
       expect.objectContaining({
@@ -1300,11 +1337,11 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     const events: unknown[] = [];
     store.subscribe((event) => events.push(event));
     await store.initialize();
-    await (store as any).runRefresh("codex");
+    await syncEngineOf(store).refresh("codex");
 
     expect(events).toEqual([
       expect.objectContaining({
@@ -1332,12 +1369,12 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     const events: unknown[] = [];
     store.subscribe((event) => events.push(event));
     await store.initialize();
     const workerCount = workerThreads.workers.length;
-    await (store as any).runRefresh("codex");
+    await syncEngineOf(store).refresh("codex");
 
     expect(core.saveCachedSessions).not.toHaveBeenCalled();
     expect(workerThreads.workers).toHaveLength(workerCount);
@@ -1383,7 +1420,7 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(true, { agents: ["codex"] });
+    const store = new LiveScanStore({ watchEnabled: true, scanOptions: { agents: ["codex"] } });
     const events: SessionsUpdatedEvent[] = [];
     store.subscribe((event) => events.push(event));
 
@@ -1456,7 +1493,7 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(true, { agents: ["codex"] });
+    const store = new LiveScanStore({ watchEnabled: true, scanOptions: { agents: ["codex"] } });
     await store.initialize();
 
     expect(fsWatch.watchers.some((watcher) => watcher.options.recursive)).toBe(false);
@@ -1512,7 +1549,7 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(true, { agents: ["codex"] });
+    const store = new LiveScanStore({ watchEnabled: true, scanOptions: { agents: ["codex"] } });
     await store.initialize();
 
     fsWatch.watchers[0]!.listener("change", join("sessions", "new.jsonl"));
@@ -1563,12 +1600,12 @@ describe("LiveScanStore", () => {
       agents: [codex, kimi],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     const events: SessionsUpdatedEvent[] = [];
     store.subscribe((event) => events.push(event));
     await store.initialize();
-    await (store as any).runRefresh("codex");
-    await (store as any).runRefresh("kimi");
+    await syncEngineOf(store).refresh("codex");
+    await syncEngineOf(store).refresh("kimi");
 
     expect(events).toEqual([]);
     await vi.advanceTimersByTimeAsync(250);
@@ -1608,17 +1645,17 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
 
     // Three schedule calls arrive within the debounce window; scheduleRefresh
     // throttles (keeps the earliest deadline) rather than debouncing, so the
     // later calls are no-ops and the run fires 200ms after the first call.
-    (store as any).scheduleRefresh("codex");
+    syncEngineOf(store).handleAgentsChanged(["codex"]);
     await vi.advanceTimersByTimeAsync(50);
-    (store as any).scheduleRefresh("codex");
+    syncEngineOf(store).handleAgentsChanged(["codex"]);
     await vi.advanceTimersByTimeAsync(50);
-    (store as any).scheduleRefresh("codex");
+    syncEngineOf(store).handleAgentsChanged(["codex"]);
     await vi.advanceTimersByTimeAsync(99);
     expect(codex.checkForChanges).not.toHaveBeenCalled();
 
@@ -1648,13 +1685,13 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
 
     // Simulate a preceding scan that took 5s: the next refresh should be
     // delayed to ~4x that cost, well beyond the 200ms debounce floor.
-    (store as any).refreshes.setLastRefreshDuration("codex", 5_000);
-    (store as any).scheduleRefresh("codex");
+    setRefreshDuration(store, "codex", 5_000);
+    syncEngineOf(store).handleAgentsChanged(["codex"]);
 
     await vi.advanceTimersByTimeAsync(19_999);
     expect(codex.checkForChanges).not.toHaveBeenCalled();
@@ -1682,16 +1719,16 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
 
-    (store as any).refreshes.setLastRefreshDuration("codex", 5_000);
+    setRefreshDuration(store, "codex", 5_000);
 
     // Writes every 3s never leave a 20s quiet gap. A plain debounce (reset on
     // every call) would push the deadline out forever; throttling on the
     // earliest deadline guarantees a refresh still fires around the 20s mark.
     for (let elapsed = 0; elapsed < 21_000; elapsed += 3_000) {
-      (store as any).scheduleRefresh("codex");
+      syncEngineOf(store).handleAgentsChanged(["codex"]);
       await vi.advanceTimersByTimeAsync(3_000);
     }
 
@@ -1714,10 +1751,10 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
 
-    (store as any).scheduleRefresh("codex", 30_000);
+    syncEngineOf(store).handleAgentsChanged(["codex"]);
     await vi.advanceTimersByTimeAsync(200);
     expect(codex.scan).not.toHaveBeenCalled();
 
@@ -1750,16 +1787,16 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
 
-    const firstRefresh = (store as any).refreshAgent("codex");
+    const firstRefresh = syncEngineOf(store).refresh("codex");
     await Promise.resolve();
     expect(checkForChanges).toHaveBeenCalledTimes(1);
 
     // A second refresh request arrives while the first is still in flight;
     // it must be deferred rather than run concurrently.
-    const secondRefresh = (store as any).refreshAgent("codex");
+    const secondRefresh = syncEngineOf(store).refresh("codex");
     await Promise.resolve();
     expect(checkForChanges).toHaveBeenCalledTimes(1);
 
@@ -1789,10 +1826,10 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
 
-    (store as any).scheduleRefresh("codex");
+    syncEngineOf(store).handleAgentsChanged(["codex"]);
     await store.shutdown();
 
     await vi.advanceTimersByTimeAsync(10_000);
@@ -1814,11 +1851,11 @@ describe("LiveScanStore", () => {
       agents: [codex],
     });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
     const listener = vi.fn();
     store.subscribe(listener);
-    const refresh = (store as any).refreshAgent("codex");
+    const refresh = syncEngineOf(store).refresh("codex");
     await vi.waitFor(() =>
       expect(workerThreads.workers.some((worker) => worker.workerData.agentName)).toBe(true),
     );
@@ -1842,7 +1879,7 @@ describe("LiveScanStore", () => {
     core.createRegisteredAgents.mockReturnValue([]);
     core.scanSessions.mockResolvedValue({ sessions: [], byAgent: {}, agents: [] });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
     workerThreads.deferSearchIndexWorkers = true;
     const job = {
@@ -1853,8 +1890,8 @@ describe("LiveScanStore", () => {
       meta: {},
     };
 
-    const current = (store as any).searchIndexJobs.enqueue("scan.refresh", [job]);
-    const pending = (store as any).searchIndexJobs.enqueue("scan.refresh", [job]);
+    const current = (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [job]);
+    const pending = (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [job]);
     const outcomes = Promise.allSettled([current, pending]);
     expect(workerThreads.workers.filter((worker) => worker.workerData.jobs)).toHaveLength(1);
 
@@ -1865,7 +1902,7 @@ describe("LiveScanStore", () => {
       expect.objectContaining({ status: "rejected", reason: expect.any(Error) }),
       expect.objectContaining({ status: "rejected", reason: expect.any(Error) }),
     ]);
-    expect((store as any).searchIndexJobs.snapshot()).toEqual(
+    expect((syncEngineOf(store) as any).searchIndexJobs.snapshot()).toEqual(
       expect.objectContaining({ activeBatchId: undefined, pendingBatches: 0 }),
     );
   });
@@ -1874,10 +1911,10 @@ describe("LiveScanStore", () => {
     core.createRegisteredAgents.mockReturnValue([]);
     core.scanSessions.mockResolvedValue({ sessions: [], byAgent: {}, agents: [] });
 
-    const store = new LiveScanStore(false, {}, {}, { deferInitialRefresh: true });
+    const store = new LiveScanStore({ watchEnabled: false, deferInitialRefresh: true });
     await store.initialize();
     workerThreads.deferSearchIndexWorkers = true;
-    const active = (store as any).searchIndexJobs.enqueue("scan.refresh", [
+    const active = (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [
       {
         kind: "full",
         context: "scan.refresh",
@@ -1887,7 +1924,7 @@ describe("LiveScanStore", () => {
       },
     ]);
     const pending = [1, 2, 3].map((version) =>
-      (store as any).searchIndexJobs.enqueue("scan.refresh", [
+      (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [
         {
           kind: "changes",
           context: "scan.refresh",
@@ -1934,7 +1971,7 @@ describe("LiveScanStore", () => {
     core.createRegisteredAgents.mockReturnValue([]);
     core.scanSessions.mockResolvedValue({ sessions: [], byAgent: {}, agents: [] });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
     workerThreads.deferSearchIndexWorkers = true;
     const job = {
@@ -1944,7 +1981,7 @@ describe("LiveScanStore", () => {
       sessions: [],
       meta: {},
     };
-    const batch = (store as any).searchIndexJobs.enqueue("scan.refresh", [job]);
+    const batch = (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [job]);
     const outcome = batch.catch((error: Error) => error);
     const worker = workerThreads.workers.at(-1)!;
 
@@ -1952,16 +1989,16 @@ describe("LiveScanStore", () => {
 
     expect(worker.terminate).toHaveBeenCalledTimes(1);
     expect(await outcome).toBeInstanceOf(Error);
-    await expect((store as any).searchIndexJobs.enqueue("scan.refresh", [job])).rejects.toThrow(
-      "Live scan store shut down",
-    );
+    await expect(
+      (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [job]),
+    ).rejects.toThrow("Live scan store shut down");
   });
 
   it("settles a worker error once when shutdown follows", async () => {
     core.createRegisteredAgents.mockReturnValue([]);
     core.scanSessions.mockResolvedValue({ sessions: [], byAgent: {}, agents: [] });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
     workerThreads.deferSearchIndexWorkers = true;
     const job = {
@@ -1971,7 +2008,7 @@ describe("LiveScanStore", () => {
       sessions: [],
       meta: {},
     };
-    const batch = (store as any).searchIndexJobs.enqueue("scan.refresh", [job]);
+    const batch = (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [job]);
     const outcome = Promise.allSettled([batch]);
     const worker = workerThreads.workers.at(-1)!;
 
@@ -1989,7 +2026,7 @@ describe("LiveScanStore", () => {
     core.createRegisteredAgents.mockReturnValue([]);
     core.scanSessions.mockResolvedValue({ sessions: [], byAgent: {}, agents: [] });
 
-    const store = new LiveScanStore(false);
+    const store = new LiveScanStore({ watchEnabled: false });
     await store.initialize();
 
     const job = {
@@ -1999,8 +2036,8 @@ describe("LiveScanStore", () => {
       sessions: [],
       meta: {},
     };
-    await (store as any).searchIndexJobs.enqueue("scan.refresh", [job]);
-    await (store as any).searchIndexJobs.enqueue("scan.refresh", [job]);
+    await (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [job]);
+    await (syncEngineOf(store) as any).searchIndexJobs.enqueue("scan.refresh", [job]);
 
     const searchIndexWorkers = workerThreads.workers.filter((worker) => worker.workerData.jobs);
     expect(searchIndexWorkers).toHaveLength(2);
